Replace Swing with JavaFX in OptionPanel
The option panel is now split into two classes: OptionPanel and OptionPanelTabLoader.
The latter is used to load in each tab group asynchronously.
For now, all controls are laid out on a GridPane with a plain Label to represent each control.
Both classes now require the use of Java 8.
The ant build file along with the class path settings will need to be modified.
The jfxrt.jar which is required for JavaFX as of now has been added.