Refactor and prepare KeyProperty
I have prepared the KeyProperty class so that it
can be properly displayed with the new Properties panel.
The key properties are currently shown as not editable labels,
but this will soon be changed to a grid of cells containing
the image icon, the label, and the associated key values.